Working on Node Project

+++ newnode/controllers/contactController.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
 const getAllContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -15,25 +15,32 @@ const getContact = asyncHandler(async (req, res) => {
 });
 
 const createContact = asyncHandler(async (req, res) => {
-  console.log("The request is : ", req.body);
   const { name, email, phone } = req.body;
   if (!name || !email || !phone) {
     res.status(400);
     throw new Error("All fields are Mandatory!");
   }
+  console.log("The request is : ", req.body);
   const contact = await Contact.create({ 
     name: name,
     email: email,
-    phone: phone
+    phone: phone,
+    //?    Ye  " user.id "  aye gi jis user ka token hum pass krein gy
+    user_id: req.user.id
   });
   res.status(201).json(contact);
 });
 
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
+  // console.log(req.user.id);
   if(!contact){
     res.status(404);
     throw new Error("Contact Not Found");
+  }
+  if(contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission");
   }
   const updateContact = await Contact.findByIdAndUpdate(
     req.params.id,
@@ -49,6 +56,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact Not Found");
   }
+   if (contact.user_id.toString() !== req.user.id) {
+     res.status(403);
+     throw new Error("User don't have permission");
+   }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
